+++ NWN.Anvil/src/main/API/Object/NwTrigger.cs
@@ -69,6 +69,7 @@ namespace Anvil.API
         if (trigger.LoadTrigger(resGff, resStruct).ToBool())
         {
           trigger.LoadObjectState(resGff, resStruct);
+          trigger.m_oidArea = Invalid;
           GC.SuppressFinalize(trigger);
           return true;
         }