+++ CHANGELOG.md
@@ -23,6 +23,7 @@ https://github.com/nwn-dotnet/Anvil/compare/v8193.34.10...HEAD
 
 ### Fixed
 - Fixed `NwGameObject.Location` setter not working for newly deserialized game objects.
+- Fixed a crash when an equipped item was sent to limbo with `NwModule.MoveObjectToLimbo`.
 
 ## 8193.34.10
 https://github.com/nwn-dotnet/Anvil/compare/v8193.34.9...v8193.34.10